+++ docs/CHANGES.rst
@@ -6,7 +6,8 @@ Changes
 2.5.3 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add parcel and applicants contents to export content [URB-2733]
+  [jchandelle]
 
 
 2.5.2 (2023-06-15)